Specify s3Bucket.quota as a "number" (instead of a "integer") (#450)
`quota` should be a u64 since it describes bytes sizes.

The Pulumi schema doesn't support u64 (or i64), so we are converting to
a float64, which should get us at least 2^52 bits of precision while
minimizing the breaking change (int -> float) for our users.

Fixes https://github.com/pulumi/pulumi-minio/issues/449

+++ sdk/java/src/main/java/com/pulumi/minio/S3Bucket.java
@@ -11,7 +11,7 @@ import com.pulumi.minio.S3BucketArgs;
 import com.pulumi.minio.Utilities;
 import com.pulumi.minio.inputs.S3BucketState;
 import java.lang.Boolean;
-import java.lang.Integer;
+import java.lang.Double;
 import java.lang.String;
 import java.util.Optional;
 import javax.annotation.Nullable;
@@ -100,10 +100,10 @@ public class S3Bucket extends com.pulumi.resources.CustomResource {
     public Output<Optional<Boolean>> objectLocking() {
         return Codegen.optional(this.objectLocking);
     }
-    @Export(name="quota", refs={Integer.class}, tree="[0]")
-    private Output</* @Nullable */ Integer> quota;
+    @Export(name="quota", refs={Double.class}, tree="[0]")
+    private Output</* @Nullable */ Double> quota;
 
-    public Output<Optional<Integer>> quota() {
+    public Output<Optional<Double>> quota() {
         return Codegen.optional(this.quota);
     }
